+++ packages/react/src/evm/providers/thirdweb-provider.tsx
@@ -165,7 +165,7 @@ export type DefaultChains = typeof defaultChains;
 
 /**
  * The `<ThirdwebProvider />` component lets you control what networks you want users to connect to,
- * what types of wallets can connect to your app, and the settings for the [Thirdweb SDK](https://docs.thirdweb.com/typescript).
+ * what types of wallets can connect to your app, and the settings for the [thirdweb SDK](https://docs.thirdweb.com/typescript).
  *
  * @example
  * You can wrap your application with the provider as follows: