+++ packages/thirdweb/src/transaction/actions/wait-for-tx-receipt.ts
@@ -12,7 +12,7 @@ const MAX_BLOCKS_WAIT_TIME = 10;
 const map = new Map<string, Promise<TransactionReceipt>>();
 
 export type WaitForReceiptOptions = TransactionOrUserOpHash & {
-  transaction: PreparedTransaction;
+  transaction: Pick<PreparedTransaction, "client" | "chain">;
 };
 
 /**